Resolves issue #9.  All functions now return POJOs, with 1 exception.
Created a "expandUrl" helper function.  This is called by the
"getExpandedUrl" and "redirectUrl" functions (which have the same
endpoint, but different media types).  This allows the shortener to be
accessed via either a web browser (for automatic redirection) or an API
client.  The "redirectUrl" function still returns a Response object, but
I felt this was cleaner than returning void but passing it an
"HttpWebResponse" object and calling the redirect on this object.